+++ sawbot-forge-1.8.9/src/main/java/dev/fivesaw/sawbot/forge/safety/SawBotMode.java
@@ -2,6 +2,5 @@ package dev.fivesaw.sawbot.forge.safety;
 
 public enum SawBotMode {
     DISABLED,
-    ENABLED,
-    FROZEN
+    ENABLED
 }